refactor: c.card move to new folder
